+++ runtime/src/transaction_payment_wrapper.rs
@@ -28,6 +28,13 @@ type RuntimeOriginOf<T> = <T as frame_system::Config>::RuntimeOrigin;
 type AccountIdOf<T> = <T as frame_system::Config>::AccountId;
 type LookupOf<T> = <T as frame_system::Config>::Lookup;
 
+/// Runtime-supplied policy: which calls have their fee charged to the signing
+/// hotkey's coldkey. Keeping the concrete list in the runtime (see `fee_filters`)
+/// lets this generic extension stay free of a hardcoded allow-list.
+pub trait ColdkeyFeeCallFilter<Call> {
+    fn charges_coldkey(call: &Call) -> bool;
+}
+
 #[freeze_struct("f003cde1f9da4a90")]
 #[derive(Encode, Decode, DecodeWithMemTracking, Clone, Eq, PartialEq, TypeInfo)]
 #[scale_info(skip_type_params(T))]
@@ -60,9 +67,8 @@ where
 impl<T: Config + pallet_proxy::Config + pallet_utility::Config + pallet_subtensor::Config>
     ChargeTransactionPaymentWrapper<T>
 where
-    RuntimeCallOf<T>: IsSubType<pallet_proxy::Call<T>>
-        + IsSubType<pallet_utility::Call<T>>
-        + IsSubType<pallet_subtensor::Call<T>>,
+    RuntimeCallOf<T>: IsSubType<pallet_proxy::Call<T>> + IsSubType<pallet_utility::Call<T>>,
+    T: ColdkeyFeeCallFilter<RuntimeCallOf<T>>,
     RuntimeOriginOf<T>: AsSystemOriginSigner<AccountIdOf<T>> + Clone,
 {
     /// Extract (real, delegate, inner_call) from a `proxy` call.
@@ -186,17 +192,21 @@ where
         common_real
     }
 
-    fn extract_coldkey_fee_payer(origin: &RuntimeOriginOf<T>) -> Option<AccountIdOf<T>> {
+    /// Determine the coldkey that should pay the fee when a hotkey is the origin.
+    ///
+    /// Returns `Some(coldkey)` only for calls the runtime marks as coldkey-paid
+    /// (via [`ColdkeyFeeCallFilter`]), and only when the signer (hotkey) has an
+    /// owner. Returns `None` otherwise, so the signer pays.
+    fn extract_coldkey_fee_payer(
+        call: &RuntimeCallOf<T>,
+        origin: &RuntimeOriginOf<T>,
+    ) -> Option<AccountIdOf<T>> {
+        if !T::charges_coldkey(call) {
+            return None;
+        }
+
         let signer = origin.as_system_origin_signer()?;
-
         pallet_subtensor::Pallet::<T>::maybe_coldkey_for_hotkey(signer)
-    }
-
-    fn is_coldkey_fee_payer_eligible(call: &RuntimeCallOf<T>) -> bool {
-        matches!(
-            call.is_sub_type(),
-            Some(pallet_subtensor::Call::set_weights { .. })
-        )
     }
 }
 
@@ -205,8 +215,8 @@ impl<T: Config + pallet_proxy::Config + pallet_utility::Config + pallet_subtenso
 where
     RuntimeCallOf<T>: Dispatchable<Info = DispatchInfo, PostInfo = PostDispatchInfo>
         + IsSubType<pallet_proxy::Call<T>>
-        + IsSubType<pallet_utility::Call<T>>
-        + IsSubType<pallet_subtensor::Call<T>>,
+        + IsSubType<pallet_utility::Call<T>>,
+    T: ColdkeyFeeCallFilter<RuntimeCallOf<T>>,
     RuntimeOriginOf<T>: AsSystemOriginSigner<AccountIdOf<T>>
         + Clone
         + From<frame_system::RawOrigin<AccountIdOf<T>>>,
@@ -247,12 +257,8 @@ where
         // Otherwise, the signer pays as usual.
         let fee_origin = if let Some(real) = Self::extract_real_fee_payer(call, &origin) {
             frame_system::RawOrigin::Signed(real).into()
-        } else if Self::is_coldkey_fee_payer_eligible(call) {
-            if let Some(coldkey) = Self::extract_coldkey_fee_payer(&origin) {
-                frame_system::RawOrigin::Signed(coldkey).into()
-            } else {
-                origin.clone()
-            }
+        } else if let Some(coldkey) = Self::extract_coldkey_fee_payer(call, &origin) {
+            frame_system::RawOrigin::Signed(coldkey).into()
         } else {
             origin.clone()
         };
@@ -332,5 +338,554 @@ where
         result: &DispatchResult,
     ) -> Result<(), TransactionValidityError> {
         ChargeTransactionPayment::<T>::bare_post_dispatch(info, post_info, len, result)
+    }
+}
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used)]
+mod tests {
+    use super::ChargeTransactionPaymentWrapper;
+    use crate::{
+        BuildStorage, NORMAL_DISPATCH_BASE_PRIORITY, OPERATIONAL_DISPATCH_PRIORITY, Proxy, Runtime,
+        RuntimeCall, RuntimeGenesisConfig, RuntimeOrigin, System, SystemCall,
+    };
+    use frame_support::{
+        assert_ok,
+        dispatch::{DispatchClass, DispatchInfo, GetDispatchInfo},
+    };
+    use pallet_subtensor_proxy as pallet_proxy;
+    use pallet_subtensor_utility as pallet_utility;
+    use pallet_transaction_payment::Val;
+    use sp_runtime::traits::{TransactionExtension, TxBaseImplication};
+    use sp_runtime::transaction_validity::{
+        TransactionSource, TransactionValidityError, ValidTransaction,
+    };
+    use subtensor_runtime_common::{AccountId, NetUid, ProxyType, TaoBalance};
+
+    const SIGNER: [u8; 32] = [1_u8; 32];
+    const REAL_A: [u8; 32] = [2_u8; 32];
+    const REAL_B: [u8; 32] = [3_u8; 32];
+    const OTHER: [u8; 32] = [4_u8; 32];
+    const BALANCE: TaoBalance = TaoBalance::new(1_000_000_000_000_u64);
+
+    fn new_test_ext() -> sp_io::TestExternalities {
+        sp_tracing::try_init_simple();
+        let mut ext: sp_io::TestExternalities = RuntimeGenesisConfig {
+            balances: pallet_balances::GenesisConfig {
+                balances: vec![
+                    (AccountId::from(SIGNER), BALANCE),
+                    (AccountId::from(REAL_A), BALANCE),
+                    (AccountId::from(REAL_B), BALANCE),
+                    (AccountId::from(OTHER), BALANCE),
+                ],
+                dev_accounts: None,
+            },
+            ..Default::default()
+        }
+        .build_storage()
+        .unwrap()
+        .into();
+        ext.execute_with(|| System::set_block_number(1));
+        ext
+    }
+
+    fn signer() -> AccountId {
+        AccountId::from(SIGNER)
+    }
+    fn real_a() -> AccountId {
+        AccountId::from(REAL_A)
+    }
+    fn real_b() -> AccountId {
+        AccountId::from(REAL_B)
+    }
+    fn other() -> AccountId {
+        AccountId::from(OTHER)
+    }
+
+    // -- Call builders --
+
+    fn call_remark() -> RuntimeCall {
+        RuntimeCall::System(SystemCall::remark {
+            remark: vec![1, 2, 3],
+        })
+    }
+
+    fn call_set_weights() -> RuntimeCall {
+        RuntimeCall::SubtensorModule(pallet_subtensor::Call::set_weights {
+            netuid: NetUid::from(1),
+            dests: vec![0],
+            weights: vec![1],
+            version_key: 0,
+        })
+    }
+
+    fn call_commit_weights() -> RuntimeCall {
+        RuntimeCall::SubtensorModule(pallet_subtensor::Call::commit_weights {
+            netuid: NetUid::from(1),
+            commit_hash: sp_core::H256::zero(),
+        })
+    }
+
+    fn proxy_call(real: AccountId, inner: RuntimeCall) -> RuntimeCall {
+        RuntimeCall::Proxy(pallet_proxy::Call::proxy {
+            real: real.into(),
+            force_proxy_type: None,
+            call: Box::new(inner),
+        })
+    }
+
+    fn proxy_announced_call(
+        delegate: AccountId,
+        real: AccountId,
+        inner: RuntimeCall,
+    ) -> RuntimeCall {
+        RuntimeCall::Proxy(pallet_proxy::Call::proxy_announced {
+            delegate: delegate.into(),
+            real: real.into(),
+            force_proxy_type: None,
+            call: Box::new(inner),
+        })
+    }
+
+    fn batch_call(calls: Vec<RuntimeCall>) -> RuntimeCall {
+        RuntimeCall::Utility(pallet_utility::Call::batch { calls })
+    }
+
+    fn batch_all_call(calls: Vec<RuntimeCall>) -> RuntimeCall {
+        RuntimeCall::Utility(pallet_utility::Call::batch_all { calls })
+    }
+
+    fn force_batch_call(calls: Vec<RuntimeCall>) -> RuntimeCall {
+        RuntimeCall::Utility(pallet_utility::Call::force_batch { calls })
+    }
+
+    // -- Setup helpers --
+
+    fn add_proxy(real: &AccountId, delegate: &AccountId) {
+        assert_ok!(Proxy::add_proxy(
+            RuntimeOrigin::signed(real.clone()),
+            delegate.clone().into(),
+            ProxyType::Any,
+            0,
+        ));
+    }
+
+    fn enable_real_pays_fee(real: &AccountId, delegate: &AccountId) {
+        assert_ok!(Proxy::set_real_pays_fee(
+            RuntimeOrigin::signed(real.clone()),
+            delegate.clone().into(),
+            true,
+        ));
+    }
+
+    // -- Validate helpers --
+
+    fn validate_call(
+        origin: RuntimeOrigin,
+        call: &RuntimeCall,
+    ) -> Result<(ValidTransaction, Val<Runtime>), TransactionValidityError> {
+        validate_call_with_info(origin, call, &call.get_dispatch_info())
+    }
+
+    fn validate_call_with_info(
+        origin: RuntimeOrigin,
+        call: &RuntimeCall,
+        info: &DispatchInfo,
+    ) -> Result<(ValidTransaction, Val<Runtime>), TransactionValidityError> {
+        let ext = ChargeTransactionPaymentWrapper::<Runtime>::new(TaoBalance::new(0));
+        let (valid_tx, val, _origin) = ext.validate(
+            origin,
+            call,
+            info,
+            100,
+            (),
+            &TxBaseImplication(()),
+            TransactionSource::External,
+        )?;
+        Ok((valid_tx, val))
+    }
+
+    /// Extract the fee payer from the validate result.
+    fn fee_payer(val: &Val<Runtime>) -> AccountId {
+        match val {
+            Val::Charge { who, .. } => who.clone(),
+            _ => panic!("expected Val::Charge"),
+        }
+    }
+
+    // ============================================================
+    // Case 0: Non-proxy calls
+    // ============================================================
+
+    #[test]
+    fn non_proxy_call_charges_signer() {
+        new_test_ext().execute_with(|| {
+            let call = call_remark();
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), signer());
+        });
+    }
+
+    // ============================================================
+    // Case 1: Simple proxy (1 level)
+    // ============================================================
+
+    #[test]
+    fn simple_proxy_charges_real_when_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_a(), &signer());
+            enable_real_pays_fee(&real_a(), &signer());
+
+            let call = proxy_call(real_a(), call_remark());
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_a());
+        });
+    }
+
+    #[test]
+    fn simple_proxy_charges_signer_when_not_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_a(), &signer());
+            // No enable_real_pays_fee
+
+            let call = proxy_call(real_a(), call_remark());
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), signer());
+        });
+    }
+
+    #[test]
+    fn proxy_announced_always_charges_signer() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+
+            // Fee propagation intentionally ignores proxy_announced; signer always pays.
+            let call = proxy_announced_call(real_b(), real_a(), call_remark());
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), signer());
+        });
+    }
+
+    // ============================================================
+    // Case 2: Nested proxy (2 levels)
+    // ============================================================
+
+    #[test]
+    fn nested_proxy_charges_inner_real_when_both_opted_in() {
+        new_test_ext().execute_with(|| {
+            // Chain: signer → real_b → real_a
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+
+            let call = proxy_call(real_b(), proxy_call(real_a(), call_remark()));
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_a());
+        });
+    }
+
+    #[test]
+    fn nested_proxy_charges_outer_real_when_only_outer_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            // No enable_real_pays_fee for A→B
+
+            let call = proxy_call(real_b(), proxy_call(real_a(), call_remark()));
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_b());
+        });
+    }
+
+    #[test]
+    fn nested_proxy_charges_signer_when_neither_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            // No enable_real_pays_fee at all
+
+            let call = proxy_call(real_b(), proxy_call(real_a(), call_remark()));
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), signer());
+        });
+    }
+
+    #[test]
+    fn nested_proxy_charges_signer_when_only_inner_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            // No enable_real_pays_fee for B→signer
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+
+            let call = proxy_call(real_b(), proxy_call(real_a(), call_remark()));
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            // Outer RealPaysFee not set → signer pays (inner opt-in is irrelevant)
+            assert_eq!(fee_payer(&val), signer());
+        });
+    }
+
+    // ============================================================
+    // Case 3: Batch of proxy calls
+    // ============================================================
+
+    #[test]
+    fn batch_charges_inner_real_when_all_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+
+            let batch = batch_call(vec![
+                proxy_call(real_a(), call_remark()),
+                proxy_call(real_a(), call_remark()),
+            ]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_a());
+        });
+    }
+
+    #[test]
+    fn batch_all_charges_inner_real_when_all_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+
+            let batch = batch_all_call(vec![
+                proxy_call(real_a(), call_remark()),
+                proxy_call(real_a(), call_remark()),
+            ]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_a());
+        });
+    }
+
+    #[test]
+    fn force_batch_charges_inner_real_when_all_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+
+            let batch = force_batch_call(vec![
+                proxy_call(real_a(), call_remark()),
+                proxy_call(real_a(), call_remark()),
+            ]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_a());
+        });
+    }
+
+    #[test]
+    fn batch_charges_outer_real_when_only_outer_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            // No enable_real_pays_fee for A→B
+
+            let batch = batch_call(vec![
+                proxy_call(real_a(), call_remark()),
+                proxy_call(real_a(), call_remark()),
+            ]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_b());
+        });
+    }
+
+    #[test]
+    fn batch_charges_outer_real_when_mixed_inner_reals() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            enable_real_pays_fee(&real_a(), &real_b());
+            add_proxy(&other(), &real_b());
+            enable_real_pays_fee(&other(), &real_b());
+
+            // Different inner reals → can't push deeper
+            let batch = batch_call(vec![
+                proxy_call(real_a(), call_remark()),
+                proxy_call(other(), call_remark()),
+            ]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_b());
+        });
+    }
+
+    #[test]
+    fn batch_charges_outer_real_when_non_proxy_in_batch() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+
+            // Batch contains a non-proxy call → extract_proxy_parts fails
+            let batch = batch_call(vec![proxy_call(real_a(), call_remark()), call_remark()]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_b());
+        });
+    }
+
+    #[test]
+    fn batch_charges_outer_real_when_empty() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+
+            let batch = batch_call(vec![]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_b());
+        });
+    }
+
+    #[test]
+    fn batch_charges_outer_real_when_inner_real_not_opted_in() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_b(), &signer());
+            enable_real_pays_fee(&real_b(), &signer());
+            add_proxy(&real_a(), &real_b());
+            // real_a has NOT opted in to pay for real_b
+
+            // Even with same real in all batch items, if RealPaysFee<A, B> not set → outer_real pays
+            let batch = batch_call(vec![
+                proxy_call(real_a(), call_remark()),
+                proxy_call(real_a(), call_remark()),
+            ]);
+            let call = proxy_call(real_b(), batch);
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(fee_payer(&val), real_b());
+        });
+    }
+
+    // ============================================================
+    // Priority override
+    // ============================================================
+
+    #[test]
+    fn priority_override_normal_dispatch() {
+        new_test_ext().execute_with(|| {
+            let call = call_remark();
+            let (valid_tx, _val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            assert_eq!(valid_tx.priority, NORMAL_DISPATCH_BASE_PRIORITY);
+        });
+    }
+
+    #[test]
+    fn priority_override_operational_dispatch() {
+        new_test_ext().execute_with(|| {
+            let call = call_remark();
+            let mut info = call.get_dispatch_info();
+            info.class = DispatchClass::Operational;
+
+            let (valid_tx, _val) =
+                validate_call_with_info(RuntimeOrigin::signed(signer()), &call, &info).unwrap();
+            assert_eq!(valid_tx.priority, OPERATIONAL_DISPATCH_PRIORITY);
+        });
+    }
+
+    #[test]
+    fn priority_override_mandatory_dispatch() {
+        new_test_ext().execute_with(|| {
+            let call = call_remark();
+            let mut info = call.get_dispatch_info();
+            info.class = DispatchClass::Mandatory;
+
+            let (valid_tx, _val) =
+                validate_call_with_info(RuntimeOrigin::signed(signer()), &call, &info).unwrap();
+            // Mandatory uses the same base as Normal
+            assert_eq!(valid_tx.priority, NORMAL_DISPATCH_BASE_PRIORITY);
+        });
+    }
+
+    #[test]
+    fn priority_override_applies_with_real_pays_fee() {
+        new_test_ext().execute_with(|| {
+            add_proxy(&real_a(), &signer());
+            enable_real_pays_fee(&real_a(), &signer());
+
+            let call = proxy_call(real_a(), call_remark());
+            let (valid_tx, _val) = validate_call(RuntimeOrigin::signed(signer()), &call).unwrap();
+            // Priority override should still apply when real pays fee
+            assert_eq!(valid_tx.priority, NORMAL_DISPATCH_BASE_PRIORITY);
+        });
+    }
+
+    // ============================================================
+    // Coldkey pays the fee when a hotkey is the origin
+    // ============================================================
+
+    #[test]
+    fn hotkey_origin_charges_coldkey_fee_payer() {
+        new_test_ext().execute_with(|| {
+            let hotkey = signer();
+            let coldkey = other();
+
+            pallet_subtensor::Owner::<Runtime>::insert(hotkey.clone(), coldkey.clone());
+
+            let call = call_set_weights();
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(hotkey), &call).unwrap();
+
+            assert_eq!(fee_payer(&val), coldkey);
+        });
+    }
+
+    // Guards against the gate only recognizing `set_weights` rather than the whole group.
+    #[test]
+    fn hotkey_origin_charges_coldkey_for_non_set_weights_member() {
+        new_test_ext().execute_with(|| {
+            let hotkey = signer();
+            let coldkey = other();
+
+            pallet_subtensor::Owner::<Runtime>::insert(hotkey.clone(), coldkey.clone());
+
+            let call = call_commit_weights();
+            let (_valid_tx, val) = validate_call(RuntimeOrigin::signed(hotkey), &call).unwrap();
+
+            assert_eq!(fee_payer(&val), coldkey);
+        });
+    }
+
+    #[test]
+    fn hotkey_origin_without_owner_charges_signer() {
+        new_test_ext().execute_with(|| {
+            let hotkey = signer();
+
+            let call = call_set_weights();
+            let (_valid_tx, val) =
+                validate_call(RuntimeOrigin::signed(hotkey.clone()), &call).unwrap();
+
+            assert_eq!(fee_payer(&val), hotkey);
+        });
+    }
+
+    // Owner is set, yet a call outside the group is never redirected: the signer pays.
+    #[test]
+    fn ineligible_call_from_hotkey_charges_signer() {
+        new_test_ext().execute_with(|| {
+            let hotkey = signer();
+            let coldkey = other();
+
+            pallet_subtensor::Owner::<Runtime>::insert(hotkey.clone(), coldkey);
+
+            let call = call_remark();
+            let (_valid_tx, val) =
+                validate_call(RuntimeOrigin::signed(hotkey.clone()), &call).unwrap();
+
+            assert_eq!(fee_payer(&val), hotkey);
+        });
     }
 }
